动态评论通知（开发中）

+++ im-platform/src/main/java/xyz/qy/implatform/service/impl/TalkStarServiceImpl.java
@@ -2,21 +2,28 @@ package xyz.qy.implatform.service.impl;
 
 import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import xyz.qy.imclient.IMClient;
 import xyz.qy.imclient.annotation.Lock;
+import xyz.qy.imcommon.model.IMTalkMessage;
+import xyz.qy.imcommon.model.IMUserInfo;
 import xyz.qy.implatform.contant.Constant;
 import xyz.qy.implatform.dto.TalkStarDTO;
 import xyz.qy.implatform.entity.CharacterAvatar;
 import xyz.qy.implatform.entity.Talk;
+import xyz.qy.implatform.entity.TalkNotify;
 import xyz.qy.implatform.entity.TalkStar;
 import xyz.qy.implatform.entity.TemplateCharacter;
 import xyz.qy.implatform.entity.User;
+import xyz.qy.implatform.enums.TalkNotifyActionTypeEnum;
 import xyz.qy.implatform.exception.GlobalException;
 import xyz.qy.implatform.mapper.TalkStarMapper;
 import xyz.qy.implatform.service.ICharacterAvatarService;
+import xyz.qy.implatform.service.ITalkNotifyService;
 import xyz.qy.implatform.service.ITalkService;
 import xyz.qy.implatform.service.ITalkStarService;
 import xyz.qy.implatform.service.ITemplateCharacterService;
@@ -24,9 +31,12 @@ import xyz.qy.implatform.service.IUserService;
 import xyz.qy.implatform.session.SessionContext;
 import xyz.qy.implatform.session.UserSession;
 import xyz.qy.implatform.util.BeanUtils;
+import xyz.qy.implatform.vo.TalkMessageVO;
 import xyz.qy.implatform.vo.TalkStarVO;
 
 import javax.annotation.Resource;
+import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.Objects;
 
 /**
@@ -47,6 +57,12 @@ public class TalkStarServiceImpl extends ServiceImpl<TalkStarMapper, TalkStar> i
 
     @Resource
     private ICharacterAvatarService characterAvatarService;
+
+    @Resource
+    private ITalkNotifyService talkNotifyService;
+
+    @Resource
+    private IMClient imClient;
 
     @Transactional
     @Lock(prefix = "im:talk:comment", key = "#talkStarDTO.getTalkId()")
@@ -99,12 +115,36 @@ public class TalkStarServiceImpl extends ServiceImpl<TalkStarMapper, TalkStar> i
         }
 
         this.save(talkStar);
+
+        if (!myUserId.equals(talk.getUserId())) {
+            TalkNotify talkNotify = new TalkNotify();
+            talkNotify.setTalkId(talkId);
+            talkNotify.setUserId(talk.getUserId());
+            talkNotify.setStarId(talkStar.getId());
+            talkNotify.setCategory(talk.getCategory());
+            talkNotify.setActionType(TalkNotifyActionTypeEnum.LIKE.getCode());
+            talkNotify.setCreateTime(LocalDateTime.now());
+            talkNotifyService.save(talkNotify);
+
+            TalkMessageVO msgInfo = new TalkMessageVO();
+            msgInfo.setType(2);
+            msgInfo.setTalk(talk);
+            msgInfo.setTalkStar(talkStar);
+
+            IMTalkMessage<TalkMessageVO> sendMessage = new IMTalkMessage<>();
+            sendMessage.setSender(new IMUserInfo(session.getUserId(), session.getTerminal()));
+            sendMessage.setRecvIds(Collections.singletonList(talk.getUserId()));
+            sendMessage.setSendResult(false);
+            sendMessage.setData(msgInfo);
+            imClient.sendTalkMessage(sendMessage);
+        }
         TalkStarVO talkStarVO = BeanUtils.copyProperties(talkStar, TalkStarVO.class);
         talkStarVO.setIsOwner(Boolean.TRUE);
         return talkStarVO;
     }
 
     @Lock(prefix = "im:talk:comment", key = "#talkStarDTO.getTalkId()")
+    @Transactional
     @Override
     public void cancelLike(TalkStarDTO talkStarDTO) {
         UserSession session = SessionContext.getSession();
@@ -112,6 +152,22 @@ public class TalkStarServiceImpl extends ServiceImpl<TalkStarMapper, TalkStar> i
         LambdaQueryWrapper<TalkStar> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(TalkStar::getTalkId, talkStarDTO.getTalkId());
         wrapper.eq(TalkStar::getUserId, myUserId);
+        wrapper.eq(TalkStar::getDeleted, false);
+
+        TalkStar talkStar = baseMapper.selectOne(wrapper);
+
+        if (ObjectUtil.isNull(talkStar)) {
+            throw new GlobalException("点赞记录不存在");
+        }
+
         baseMapper.delete(wrapper);
+        
+        // 删除点赞提醒记录
+        LambdaUpdateWrapper<TalkNotify> updateWrapper = new LambdaUpdateWrapper<>();
+        updateWrapper.eq(TalkNotify::getStarId, talkStar.getId());
+        updateWrapper.eq(TalkNotify::getActionType, TalkNotifyActionTypeEnum.LIKE.getCode());
+        updateWrapper.eq(TalkNotify::getTalkId, talkStarDTO.getTalkId());
+
+        talkNotifyService.remove(updateWrapper);
     }
 }
